consolidate client-handling code; makefile tweaks
- moved 'ping' function to client.go
- added gowatch.yml to exclude directories on build
