+++ gowatch.yml
@@ -1,0 +1,13 @@
+# main executable package
+build_pkg: ./cmd
+
+# path to build
+output: dist/pickaxx_dev
+
+# exclude changes in these paths from triggering build
+excluded_paths:
+    - testserver
+    - node_modules
+    - dist
+    - public
+    - templates
